backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
aiofiles>=23.2.1
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
websockets>=12.0
